moved notebook to docs/example, deleted not needed utils.py, moved chassis loop to experimental folder, updated __init__.py
